backend/requirements.txt
yara-python==4.5.1
fastapi
uvicorn
websockets
